Remove extra text component, Move homepage to page
