--- client/src/components/ChapterExtraText.tsx
@@ -1,29 +1,0 @@
-const ChapterExtraText = () => (
-  <div>
-    <ul className="space-y-2">
-      <li>
-        <strong>Plastic:</strong> materiaal gemaakt van fossiele brandstoffen. Duurt honderden jaren voordat het afbreekt.
-      </li>
-      <li>
-        <strong>Verplaatsing:</strong> plastic kan direct in de oceaan worden gegooid of indirect door rivieren, wind en golven.
-      </li>
-      <li>
-        <strong>Microplastics:</strong> piepkleine stukjes plastic die slecht zijn voor de gezondheid van allerlei organismen.
-      </li>
-      <li>
-        <strong>Schade van plastic:</strong> plasticvervuiling brengt schade aan dieren, onze gezondheid, de economie en het klimaat.
-      </li>
-      <li>
-        <strong>Klimaatverandering:</strong> klimaatverandering is het langzaam warmer worden van de aarde door vervuiling, waardoor het weer extremer wordt en dieren en mensen in de problemen komen.
-      </li>
-      <li>
-        <strong>Broeikasgassen:</strong> gassen in de atmosfeer. Ze laten de zonnestralen binnen, maar houden een deel van de warmte vast, waardoor het warmer wordt op onze planeet.
-      </li>
-      <li>
-        <strong>Opruimacties:</strong> zwerfvuil opruimen in je lokale gemeenschap.
-      </li>
-    </ul>
-  </div>
-);
-
-export default ChapterExtraText;
